refactored global vars/functions
